move everything SAML into samlmd.py and do lazy registration of parser to avoid loop
