+++ src/pyff/samlmd.py
@@ -14,7 +14,8 @@ from .exceptions import *
 import traceback
 from six import StringIO
 from requests import ConnectionError
-
+from .fetch import ResourceManager
+from .parse import add_parser
 
 class EntitySet(object):
     def __init__(self, initial=None):
@@ -123,7 +124,7 @@ def parse_saml_metadata(source,
     return t, expire_time_offset
 
 
-class SAMLMetadataResourceParser():
+class SAMLMetadataResourceParser:
     def __init__(self):
         pass
 
@@ -153,6 +154,7 @@ class SAMLMetadataResourceParser():
 
         return info
 
+add_parser(SAMLMetadataResourceParser())
 
 def metadata_expiration(t):
     relt = root(t)
@@ -180,7 +182,7 @@ def filter_invalids_from_document(t, base_url, validation_errors):
         if not xsd.validate(e):
             log.debug(etree.tostring(e))
             error = xml_error(xsd.error_log, m=base_url)
-            entity_id = e.get("entityID","(Missing entityID)")
+            entity_id = e.get("entityID", "(Missing entityID)")
             log.warn('removing \'%s\': schema validation failed: %s' % (entity_id, xsd.error_log))
             validation_errors[entity_id] = "{}".format(xsd.error_log)
             if e.getparent() is None:
@@ -274,7 +276,7 @@ Produce an EntityDescriptors set from a list of entities. Optional Name, cacheDu
                                source="request",
                                validation_errors=validation_errors)
 
-        for base_url,err in validation_errors.items():
+        for base_url, err in validation_errors.items():
             log.error("Validation error: @ {}: {}".format(base_url, err))
 
     return t
@@ -462,6 +464,7 @@ def entity_attribute_dict(entity):
         d[ATTRS['software']] = [guess_entity_software(entity)]
 
     return d
+
 
 def gen_icon(e):
     scopes = entity_scopes(e)
@@ -662,6 +665,7 @@ def entity_simple_summary(e):
         d['privacy_statement_url'] = psu
 
     return d
+
 
 def entity_orgurl(entity, langs=None):
     for organizationUrl in filter_lang(entity.iter("{%s}OrganizationURL" % NS['md']), langs=langs):
@@ -915,6 +919,7 @@ Entities where no value exists for the given 'sxp' are sorted last.
 :param t: An element tree containing the entities to sort
 :param sxp: xpath expression selecting the value used for sorting the entities
 """
+
     def get_key(e):
         eid = e.attrib.get('entityID')
         sv = None
@@ -937,3 +942,79 @@ Entities where no value exists for the given 'sxp' are sorted last.
 
     container = root(t)
     container[:] = sorted(container, key=lambda e: get_key(e))
+
+
+class MDRepository():
+    """A class representing a set of SAML metadata and the resources from where this metadata was loaded.
+    """
+
+    def __init__(self):
+        self.rm = ResourceManager()
+        self.store = None
+
+    def _lookup(self, member, store=None):
+        if store is None:
+            store = self.store
+
+        if member is None:
+            member = "entities"
+
+        if type(member) is str or type(member) is unicode:
+            if '!' in member:
+                (src, xp) = member.split("!")
+                if len(src) == 0:
+                    src = None
+                return self.lookup(src, xp=xp, store=store)
+
+        log.debug("calling store lookup %s" % member)
+        return store.lookup(member)
+
+    def lookup(self, member, xp=None, store=None):
+        """
+Lookup elements in the working metadata repository
+
+:param member: A selector (cf below)
+:type member: basestring
+:param xp: An optional xpath filter
+:type xp: basestring
+:param store: the store to operate on
+:return: An interable of EntityDescriptor elements
+:rtype: etree.Element
+
+
+**Selector Syntax**
+
+    - selector "+" selector
+    - [sourceID] "!" xpath
+    - attribute=value or {attribute}value
+    - entityID
+    - source (typically @Name from an EntitiesDescriptor set but could also be an alias)
+
+The first form results in the intersection of the results of doing a lookup on the selectors. The second form
+results in the EntityDescriptor elements from the source (defaults to all EntityDescriptors) that match the
+xpath expression. The attribute-value forms resuls in the EntityDescriptors that contain the specified entity
+attribute pair. If non of these forms apply, the lookup is done using either source ID (normally @Name from
+the EntitiesDescriptor) or the entityID of single EntityDescriptors. If member is a URI but isn't part of
+the metadata repository then it is fetched an treated as a list of (one per line) of selectors. If all else
+fails an empty list is returned.
+
+        """
+        if store is None:
+            store = self.store
+
+        l = self._lookup(member, store=store)
+        if hasattr(l, 'tag'):
+            l = [l]
+        elif hasattr(l, '__iter__'):
+            l = list(l)
+
+        if xp is None:
+            return l
+        else:
+            log.debug("filtering %d entities using xpath %s" % (len(l), xp))
+            t = entitiesdescriptor(l, 'dummy', lookup_fn=self.lookup)
+            if t is None:
+                return []
+            l = root(t).xpath(xp, namespaces=NS, smart_strings=False)
+            log.debug("got %d entities after filtering" % len(l))
+            return l
